+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/fagsak/FagsakService.kt
@@ -57,8 +57,6 @@ class FagsakService(
     private val personidentService: PersonidentService,
     private val utvidetBehandlingService: UtvidetBehandlingService,
     private val behandlingService: BehandlingService,
-    private val personopplysningerService: PersonopplysningerService,
-    private val familieIntegrasjonerTilgangskontrollService: FamilieIntegrasjonerTilgangskontrollService,
     private val saksstatistikkEventPublisher: SaksstatistikkEventPublisher,
     private val skyggesakService: SkyggesakService,
     private val vedtaksperiodeService: VedtaksperiodeService,
@@ -67,7 +65,6 @@ class FagsakService(
     private val behandlingHentOgPersisterService: BehandlingHentOgPersisterService,
     private val skjermetBarnSøkerRepository: SkjermetBarnSøkerRepository,
     private val unleashService: UnleashNextMedContextService,
-    private val integrasjonClient: IntegrasjonClient,
 ) {
     private val antallFagsakerOpprettetFraManuell =
         Metrics.counter("familie.ba.sak.fagsak.opprettet", "saksbehandling", "manuell")
@@ -317,206 +314,6 @@ class FagsakService(
 
     fun hentLøpendeFagsaker(): List<Fagsak> = fagsakRepository.finnLøpendeFagsaker()
 
-    fun hentFagsakDeltager(personIdent: String): List<RestFagsakDeltager> {
-        val aktør = personidentService.hentAktørOrNullHvisIkkeAktivFødselsnummer(personIdent) ?: return emptyList()
-
-        val maskertDeltaker =
-            runCatching {
-                hentMaskertFagsakdeltakerVedManglendeTilgang(aktør)
-            }.fold(
-                onSuccess = { it },
-                onFailure = { return sjekkStatuskodeOgHåndterFeil(it) },
-            )
-
-        if (maskertDeltaker != null) {
-            return listOf(maskertDeltaker)
-        }
-
-        val personInfoMedRelasjoner =
-            runCatching {
-                personopplysningerService.hentPersoninfoMedRelasjonerOgRegisterinformasjon(aktør)
-            }.fold(
-                onSuccess = { it },
-                onFailure = { return sjekkStatuskodeOgHåndterFeil(it) },
-            )
-        val assosierteFagsakDeltagere = hentAssosierteFagsakdeltagere(aktør, personInfoMedRelasjoner)
-
-        val erBarn = Period.between(personInfoMedRelasjoner.fødselsdato, LocalDate.now()).years < 18
-
-        val fagsaker = fagsakRepository.finnFagsakerForAktør(aktør).ifEmpty { listOf(null) }
-        fagsaker.forEach { fagsak ->
-            if (assosierteFagsakDeltagere.find { it.ident == aktør.aktivFødselsnummer() && it.fagsakId == fagsak?.id } == null) {
-                assosierteFagsakDeltagere.add(
-                    RestFagsakDeltager(
-                        navn = personInfoMedRelasjoner.navn,
-                        ident = aktør.aktivFødselsnummer(),
-                        // we set the role to unknown when the person is not a child because the person may not have a child
-                        rolle = if (erBarn) FagsakDeltagerRolle.BARN else FagsakDeltagerRolle.UKJENT,
-                        kjønn = personInfoMedRelasjoner.kjønn,
-                        fagsakId = fagsak?.id,
-                        fagsakType = fagsak?.type,
-                        adressebeskyttelseGradering = personInfoMedRelasjoner.adressebeskyttelseGradering,
-                    ),
-                )
-            }
-        }
-
-        if (erBarn) {
-            personInfoMedRelasjoner.forelderBarnRelasjon
-                .filter { relasjon ->
-                    relasjon.relasjonsrolle == FORELDERBARNRELASJONROLLE.FAR ||
-                        relasjon.relasjonsrolle == FORELDERBARNRELASJONROLLE.MOR ||
-                        relasjon.relasjonsrolle == FORELDERBARNRELASJONROLLE.MEDMOR
-                }.forEach { relasjon ->
-                    if (assosierteFagsakDeltagere.find { fagsakDeltager ->
-                            fagsakDeltager.ident == relasjon.aktør.aktivFødselsnummer()
-                        } == null
-                    ) {
-                        val maskertForelder =
-                            hentMaskertFagsakdeltakerVedManglendeTilgang(relasjon.aktør)
-                        if (maskertForelder != null) {
-                            assosierteFagsakDeltagere.add(maskertForelder.copy(rolle = FagsakDeltagerRolle.FORELDER))
-                        } else {
-                            val fagsakerForRelasjon = fagsakRepository.finnFagsakerForAktør(relasjon.aktør).ifEmpty { listOf(null) }
-                            fagsakerForRelasjon.forEach { fagsak ->
-                                assosierteFagsakDeltagere.add(
-                                    RestFagsakDeltager(
-                                        navn = relasjon.navn,
-                                        ident = relasjon.aktør.aktivFødselsnummer(),
-                                        rolle = FagsakDeltagerRolle.FORELDER,
-                                        kjønn = relasjon.kjønn,
-                                        fagsakId = fagsak?.id,
-                                        fagsakType = fagsak?.type,
-                                        adressebeskyttelseGradering = relasjon.adressebeskyttelseGradering,
-                                    ),
-                                )
-                            }
-                        }
-                    }
-                }
-        }
-        val fagsakDeltagereMedEgenAnsattStatus = settEgenAnsattStatusPåFagsakDeltagere(assosierteFagsakDeltagere)
-
-        return fagsakDeltagereMedEgenAnsattStatus
-    }
-
-    private fun sjekkStatuskodeOgHåndterFeil(throwable: Throwable): List<RestFagsakDeltager> {
-        val clientError = throwable as? HttpStatusCodeException?
-        return if ((clientError != null && clientError.statusCode == HttpStatus.NOT_FOUND) ||
-            throwable.message?.contains("Fant ikke person") == true
-        ) {
-            emptyList()
-        } else {
-            throw throwable
-        }
-    }
-
-    private fun hentAssosierteFagsakdeltagere(
-        aktør: Aktør,
-        personInfoMedRelasjoner: PersonInfo,
-    ): MutableList<RestFagsakDeltager> {
-        val assosierteFagsakDeltagerMap = mutableMapOf<Long, RestFagsakDeltager>()
-
-        personRepository.findByAktør(aktør).forEach { person: Person ->
-            if (!person.personopplysningGrunnlag.aktiv) {
-                return@forEach
-            }
-            val behandling = behandlingHentOgPersisterService.hent(behandlingId = person.personopplysningGrunnlag.behandlingId)
-            if (!behandling.aktiv || behandling.fagsak.arkivert || assosierteFagsakDeltagerMap.containsKey(behandling.fagsak.id)) {
-                return@forEach
-            }
-
-            assosierteFagsakDeltagerMap[behandling.fagsak.id] = hentAssosiertFagsakdeltager(behandling, aktør, personInfoMedRelasjoner)
-        }
-
-        return assosierteFagsakDeltagerMap.values.toMutableList()
-    }
-
-    private fun hentAssosiertFagsakdeltager(
-        behandling: Behandling,
-        aktør: Aktør,
-        personInfoMedRelasjoner: PersonInfo,
-    ): RestFagsakDeltager {
-        if (behandling.fagsak.aktør == aktør) {
-            return RestFagsakDeltager(
-                navn = personInfoMedRelasjoner.navn,
-                ident = behandling.fagsak.aktør.aktivFødselsnummer(),
-                rolle =
-                    when (behandling.fagsak.type) {
-                        FagsakType.NORMAL -> FagsakDeltagerRolle.FORELDER
-                        FagsakType.SKJERMET_BARN -> FagsakDeltagerRolle.BARN
-                        FagsakType.BARN_ENSLIG_MINDREÅRIG -> FagsakDeltagerRolle.BARN
-                        FagsakType.INSTITUSJON -> FagsakDeltagerRolle.UKJENT
-                    },
-                kjønn = personInfoMedRelasjoner.kjønn,
-                fagsakId = behandling.fagsak.id,
-                fagsakType = behandling.fagsak.type,
-                adressebeskyttelseGradering = personInfoMedRelasjoner.adressebeskyttelseGradering,
-            )
-        }
-
-        val maskertForelder = hentMaskertFagsakdeltakerVedManglendeTilgang(behandling.fagsak.aktør)
-        if (maskertForelder != null) {
-            return maskertForelder.copy(
-                rolle = FagsakDeltagerRolle.FORELDER,
-                fagsakType = behandling.fagsak.type,
-            )
-        }
-
-        val forelderInfo = personInfoMedRelasjoner.forelderBarnRelasjon.find { it.aktør.aktivFødselsnummer() == behandling.fagsak.aktør.aktivFødselsnummer() }
-        if (forelderInfo != null) {
-            return RestFagsakDeltager(
-                navn = forelderInfo.navn,
-                ident = behandling.fagsak.aktør.aktivFødselsnummer(),
-                rolle = FagsakDeltagerRolle.FORELDER,
-                kjønn = forelderInfo.kjønn,
-                fagsakId = behandling.fagsak.id,
-                fagsakType = behandling.fagsak.type,
-                adressebeskyttelseGradering = forelderInfo.adressebeskyttelseGradering,
-            )
-        }
-
-        // Person med forelderrolle uten direkte relasjon
-        return runCatching {
-            personopplysningerService.hentPersoninfoEnkel(behandling.fagsak.aktør)
-        }.fold(
-            onSuccess = {
-                RestFagsakDeltager(
-                    navn = it.navn,
-                    ident = behandling.fagsak.aktør.aktivFødselsnummer(),
-                    rolle = FagsakDeltagerRolle.FORELDER,
-                    kjønn = it.kjønn,
-                    fagsakId = behandling.fagsak.id,
-                    fagsakType = behandling.fagsak.type,
-                    adressebeskyttelseGradering = it.adressebeskyttelseGradering,
-                )
-            },
-            onFailure = {
-                throw Feil("Feil ved henting av person fra PDL", throwable = it)
-            },
-        )
-    }
-
-    private fun hentMaskertFagsakdeltakerVedManglendeTilgang(aktør: Aktør): RestFagsakDeltager? =
-        familieIntegrasjonerTilgangskontrollService
-            .hentMaskertPersonInfoVedManglendeTilgang(aktør)
-            ?.let {
-                RestFagsakDeltager(
-                    rolle = FagsakDeltagerRolle.UKJENT,
-                    adressebeskyttelseGradering = it.adressebeskyttelseGradering,
-                    harTilgang = false,
-                )
-            }
-
-    private fun settEgenAnsattStatusPåFagsakDeltagere(fagsakDeltagere: MutableList<RestFagsakDeltager>): List<RestFagsakDeltager> {
-        val egenAnsattPerIdent = integrasjonClient.sjekkErEgenAnsattBulk(fagsakDeltagere.map { it.ident })
-        return fagsakDeltagere.map { fagsakDeltager ->
-            fagsakDeltager.copy(
-                erEgenAnsatt = egenAnsattPerIdent.getOrDefault(fagsakDeltager.ident, null),
-            )
-        }
-    }
-
     fun finnAlleFagsakerHvorAktørHarLøpendeYtelseAvType(
         aktør: Aktør,
         ytelseTyper: List<YtelseType>,
@@ -545,39 +342,6 @@ class FagsakService(
         val fagsakerHvorAktørHarLøpendeOrdinærBarnetrygd = finnAlleFagsakerHvorAktørHarLøpendeYtelseAvType(aktør = aktør, ytelseTyper = listOf(YtelseType.ORDINÆR_BARNETRYGD))
 
         return (alleLøpendeFagsakerPåAktør + fagsakerHvorAktørHarLøpendeOrdinærBarnetrygd).distinct()
-    }
-
-    fun oppgiFagsakdeltagere(
-        aktør: Aktør,
-        barnasAktørId: List<Aktør>,
-    ): List<RestFagsakDeltager> {
-        val fagsakDeltagere = mutableListOf<RestFagsakDeltager>()
-
-        hentFagsakPåPerson(aktør)?.also { fagsak ->
-            fagsakDeltagere.add(
-                RestFagsakDeltager(
-                    ident = aktør.aktivFødselsnummer(),
-                    fagsakId = fagsak.id,
-                    fagsakStatus = fagsak.status,
-                    rolle = FagsakDeltagerRolle.FORELDER,
-                ),
-            )
-        }
-
-        barnasAktørId.forEach { barnsAktørId ->
-            hentFagsakerPåPerson(barnsAktørId).toSet().forEach { fagsak ->
-                fagsakDeltagere.add(
-                    RestFagsakDeltager(
-                        ident = barnsAktørId.aktivFødselsnummer(),
-                        fagsakId = fagsak.id,
-                        fagsakStatus = fagsak.status,
-                        rolle = FagsakDeltagerRolle.BARN,
-                    ),
-                )
-            }
-        }
-
-        return fagsakDeltagere
     }
 
     companion object {
