+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/kjerne/fagsak/FagsakServiceIntegrationTest.kt
@@ -85,260 +85,6 @@ class FagsakServiceIntegrationTest(
         databaseCleanupService.truncate()
     }
 
-    /*
-    This is a complicated test against following family relationship:
-    søker3-----------
-    (no case)       | (medmor)
-                    barn2
-                    | (medmor)
-    søker1-----------
-                    | (mor)
-                    barn1
-                    | (far)
-    søker2-----------
-                    | (far)
-                    barn3
-
-     We tests three search:
-     1) search for søker1, one participant (søker1) should be returned
-     2) search for barn1, three participants (barn1, søker1, søker2) should be returned
-     3) search for barn2, three participants (barn2, søker3, søker1) should be returned, where fagsakId of søker3 is null
-     */
-    @Test
-    fun `test å søke fagsak med fnr`() {
-        val søker1Fnr = randomFnr()
-        val søker2Fnr = randomFnr()
-        val søker3Fnr = randomFnr()
-        val barn1Fnr = randomFnr()
-        val barn2Fnr = randomFnr()
-        val barn3Fnr = randomFnr()
-
-        val søker1Aktør = personidentService.hentAktør(søker1Fnr)
-        val søker3Aktør = personidentService.hentAktør(søker3Fnr)
-
-        leggTilPersonInfo(
-            barn1Fnr,
-            PersonInfo(fødselsdato = LocalDate.of(2018, 5, 1), kjønn = Kjønn.KVINNE, navn = "barn1"),
-        )
-
-        leggTilPersonInfo(
-            barn2Fnr,
-            PersonInfo(
-                fødselsdato = LocalDate.of(2019, 5, 1),
-                kjønn = Kjønn.MANN,
-                navn = "barn2",
-                forelderBarnRelasjon =
-                    setOf(
-                        ForelderBarnRelasjon(
-                            søker1Aktør,
-                            FORELDERBARNRELASJONROLLE.MEDMOR,
-                            "søker1",
-                            LocalDate.of(1990, 2, 19),
-                        ),
-                        ForelderBarnRelasjon(
-                            søker3Aktør,
-                            FORELDERBARNRELASJONROLLE.MEDMOR,
-                            "søker3",
-                            LocalDate.of(1990, 1, 10),
-                        ),
-                    ),
-            ),
-        )
-
-        leggTilPersonInfo(
-            søker1Fnr,
-            PersonInfo(fødselsdato = LocalDate.of(1990, 2, 19), kjønn = Kjønn.KVINNE, navn = "søker1"),
-        )
-
-        leggTilPersonInfo(
-            søker2Fnr,
-            PersonInfo(fødselsdato = LocalDate.of(1991, 2, 20), kjønn = Kjønn.MANN, navn = "søker2"),
-        )
-
-        leggTilPersonInfo(
-            barn3Fnr,
-            PersonInfo(fødselsdato = LocalDate.of(2017, 3, 1), kjønn = Kjønn.KVINNE, navn = "barn3"),
-        )
-
-        leggTilPersonInfo(
-            søker3Fnr,
-            PersonInfo(fødselsdato = LocalDate.of(1990, 1, 10), kjønn = Kjønn.KVINNE, navn = "søker3"),
-        )
-
-        val fagsak0 =
-            fagsakService.hentEllerOpprettFagsak(
-                FagsakRequest(
-                    søker1Fnr,
-                ),
-            )
-
-        val fagsak1 =
-            fagsakService.hentEllerOpprettFagsak(
-                FagsakRequest(
-                    søker2Fnr,
-                ),
-            )
-
-        val førsteBehandling =
-            stegService.håndterNyBehandling(
-                NyBehandling(
-                    kategori = BehandlingKategori.NASJONAL,
-                    underkategori = BehandlingUnderkategori.ORDINÆR,
-                    søkersIdent = søker1Fnr,
-                    behandlingType = BehandlingType.FØRSTEGANGSBEHANDLING,
-                    søknadMottattDato = LocalDate.now(),
-                    fagsakId = fagsak0.data!!.id,
-                ),
-            )
-        stegService.håndterPersongrunnlag(
-            førsteBehandling,
-            RegistrerPersongrunnlagDTO(ident = søker1Fnr, barnasIdenter = listOf(barn1Fnr)),
-        )
-
-        behandlingService.oppdaterStatusPåBehandling(førsteBehandling.id, BehandlingStatus.AVSLUTTET)
-
-        val andreBehandling =
-            stegService.håndterNyBehandling(
-                NyBehandling(
-                    kategori = BehandlingKategori.NASJONAL,
-                    underkategori = BehandlingUnderkategori.ORDINÆR,
-                    søkersIdent = søker1Fnr,
-                    behandlingType = BehandlingType.FØRSTEGANGSBEHANDLING,
-                    søknadMottattDato = LocalDate.now(),
-                    fagsakId = fagsak0.data!!.id,
-                ),
-            )
-        stegService.håndterPersongrunnlag(
-            andreBehandling,
-            RegistrerPersongrunnlagDTO(
-                ident = søker1Fnr,
-                barnasIdenter = listOf(barn1Fnr, barn2Fnr),
-            ),
-        )
-
-        val tredjeBehandling =
-            stegService.håndterNyBehandling(
-                NyBehandling(
-                    kategori = BehandlingKategori.NASJONAL,
-                    underkategori = BehandlingUnderkategori.ORDINÆR,
-                    søkersIdent = søker2Fnr,
-                    behandlingType = BehandlingType.FØRSTEGANGSBEHANDLING,
-                    søknadMottattDato = LocalDate.now(),
-                    fagsakId = fagsak1.data!!.id,
-                ),
-            )
-        stegService.håndterPersongrunnlag(
-            tredjeBehandling,
-            RegistrerPersongrunnlagDTO(ident = søker2Fnr, barnasIdenter = listOf(barn1Fnr)),
-        )
-
-        val søkeresultat1 = fagsakService.hentFagsakDeltager(søker1Fnr)
-        assertEquals(1, søkeresultat1.size)
-        assertEquals(Kjønn.KVINNE, søkeresultat1[0].kjønn)
-        assertEquals("søker1", søkeresultat1[0].navn)
-        assertEquals(fagsak0.data!!.id, søkeresultat1[0].fagsakId)
-
-        val søkeresultat2 = fagsakService.hentFagsakDeltager(barn1Fnr)
-        assertEquals(3, søkeresultat2.size)
-        var matching = 0
-        søkeresultat2.forEach {
-            matching +=
-                if (it.fagsakId == fagsak0.data!!.id) {
-                    1
-                } else if (it.fagsakId == fagsak1.data!!.id) {
-                    10
-                } else {
-                    0
-                }
-        }
-        assertEquals(11, matching)
-        assertEquals(1, søkeresultat2.filter { it.ident == barn1Fnr }.size)
-
-        val søkeresultat3 = fagsakService.hentFagsakDeltager(barn2Fnr)
-        assertEquals(3, søkeresultat3.size)
-        assertEquals(1, søkeresultat3.filter { it.ident == barn2Fnr }.size)
-        assertNull(søkeresultat3.find { it.ident == barn2Fnr }!!.fagsakId)
-        assertEquals(fagsak0.data!!.id, søkeresultat3.find { it.ident == søker1Fnr }!!.fagsakId)
-        assertEquals(1, søkeresultat3.filter { it.ident == søker3Fnr }.size)
-        assertEquals("søker3", søkeresultat3.filter { it.ident == søker3Fnr }[0].navn)
-        assertNull(søkeresultat3.find { it.ident == søker3Fnr }!!.fagsakId)
-
-        val fagsak = fagsakService.hentNormalFagsak(søker1Aktør)!!
-
-        assertEquals(
-            FagsakStatus.OPPRETTET.name,
-            saksstatistikkMellomlagringRepository
-                .findByTypeAndTypeId(SAK, fagsak.id)
-                .last()
-                .jsonToSakDVH()
-                .sakStatus,
-        )
-    }
-
-    @Test
-    fun `Skal teste at arkiverte fagsaker med behandling ikke blir funnet ved søk`() {
-        val søker1Fnr = randomFnr()
-        val søker1Aktør = tilAktør(søker1Fnr)
-
-        leggTilPersonInfo(
-            søker1Fnr,
-            PersonInfo(fødselsdato = LocalDate.of(1991, 2, 19), kjønn = Kjønn.KVINNE, navn = "søker1"),
-        )
-
-        val fagsak =
-            fagsakService.hentEllerOpprettFagsak(
-                FagsakRequest(
-                    søker1Fnr,
-                ),
-            )
-
-        stegService.håndterNyBehandling(
-            NyBehandling(
-                kategori = BehandlingKategori.NASJONAL,
-                underkategori = BehandlingUnderkategori.ORDINÆR,
-                søkersIdent = søker1Fnr,
-                behandlingType = BehandlingType.FØRSTEGANGSBEHANDLING,
-                søknadMottattDato = LocalDate.now(),
-                fagsakId = fagsak.data!!.id,
-            ),
-        )
-
-        fagsakService.lagre(
-            fagsakService.hentFagsakPåPerson(søker1Aktør).also { it?.arkivert = true }!!,
-        )
-
-        val søkeresultat1 = fagsakService.hentFagsakDeltager(søker1Fnr)
-
-        assertEquals(1, søkeresultat1.size)
-        assertNull(søkeresultat1.first().fagsakId)
-    }
-
-    @Test
-    fun `Skal teste at arkiverte fagsaker uten behandling ikke blir funnet ved søk`() {
-        val søker1Fnr = randomFnr()
-        val søker1Aktør = tilAktør(søker1Fnr)
-
-        leggTilPersonInfo(
-            søker1Fnr,
-            PersonInfo(fødselsdato = LocalDate.of(1992, 2, 19), kjønn = Kjønn.KVINNE, navn = "søker1"),
-        )
-
-        fagsakService.hentEllerOpprettFagsak(
-            FagsakRequest(
-                søker1Fnr,
-            ),
-        )
-
-        fagsakService.lagre(
-            fagsakService.hentFagsakPåPerson(søker1Aktør).also { it?.arkivert = true }!!,
-        )
-
-        val søkeresultat1 = fagsakService.hentFagsakDeltager(søker1Fnr)
-
-        assertEquals(1, søkeresultat1.size)
-        assertNull(søkeresultat1.first().fagsakId)
-    }
-
     @Test
     fun `Skal teste at man henter alle fagsakene til barnet`() {
         val barnFnr = randomFnr()
@@ -369,33 +115,6 @@ class FagsakServiceIntegrationTest(
         val fagsaker = fagsakService.hentFagsakerPåPerson(barnAktør.first())
         assertEquals(2, fagsaker.size)
         assertThat(persongrunnlagRepository.findAll()).hasSize(4)
-    }
-
-    // Satte XX for at dette testet skal kjøre sist.
-    @Test
-    fun `XX Søk på fnr som ikke finnes i PDL skal vi tom liste`() {
-        every {
-            mockFamilieIntegrasjonerTilgangskontrollClient.sjekkTilgangTilPersoner(any())
-        } answers {
-            throw HttpServerErrorException(
-                HttpStatus.INTERNAL_SERVER_ERROR,
-                "[PdlRestClient][Feil ved oppslag på person: Fant ikke person]",
-            )
-        }
-        assertEquals(emptyList<RestFagsakDeltager>(), fagsakService.hentFagsakDeltager(randomFnr()))
-    }
-
-    @Test
-    fun `Skal returnere tom liste ved søk hvis ident ikke har aktiv fødselsnummer`() {
-        val fnr = randomFnr()
-        fakePdlIdentRestClient.leggTilIdent(
-            fnr,
-            listOf(
-                IdentInformasjon("npid", gruppe = "NPID", historisk = false),
-                IdentInformasjon("122334343", gruppe = "AKTOERID", historisk = false),
-            ),
-        )
-        assertThat(fagsakService.hentFagsakDeltager(fnr)).hasSize(0)
     }
 
     @Test
